moved WithState from helpers to client_api
